grant @jivan's wish (move the task tests to `test_tasks`)
